Fix apical point detection
Sometimes TMD find apical point distance from soma that
is greater than the last bifurcation distance, so TNS
can not detect the last bifurcation and thus return a
apical_point = None, which can cause issues for other codes.
Now TNS always returns an apical_point, either the same as
before (detected using the distance given by the TMD code)
or the last bifurcation if the apical point was not detected.

Change-Id: I50df99e8e4bc9e5aad2f6e87801ba4be59826401