Refactor and introduce orientation manager
Orientation manager takes care of the orientations for each tree
type given a strict mode and values from the config. For each mode
a new method must be implemented and all methods are decoupled from
each other. However, the implementation of the orientation manager
allows to take into account for each step the orientations of all
previous steps if avoidances needs to be implemented.

* orientations.py was added for the orientation manager and its related
  functions
* A soma data structure was added that decoupled the soma
  data from the soma_grower mutation methods
* soma_grower and soma used the orientations.py functions to ensure
  that new and old implementations are identical
* A legacy switch is added to allow old configs to still work
* Astrocyte math_utils were merged with the general morphmath.utils
* point_array from astrocyte directory was moved to morphmath dir

WIP: functional tests are added one by one to make sure everything works
exactly the same as before.

Change-Id: If42b65e70f96b8089e412ce8ef53a2c422c87a36