Fix edge case for astrocyte target grower
In the case of an opposite initial direction away from the target
and for a mixing factor p=0.5 the total direction was zero. When
normalized that led to inf and and thus an infinite loop. This was
fixed by explicitly checking if the total direction is zero and
skipping normalization and point update in that case.

Change-Id: Id3afb50b483b4b55d4bd5b3f2d419ad635055381